reduce duplication and refactor
