+++ webview/src/shared/hooks/useVsCodeMessaging.ts
@@ -3,15 +3,31 @@ import {
   MessageToWebview,
   WebviewData
 } from 'dvc/src/webview/contract'
-import { useEffect } from 'react'
+import { useCallback, useEffect } from 'react'
+import { useDispatch } from 'react-redux'
 import { sendMessage } from '../vscode'
+import { ExperimentsDispatch } from '../../experiments/store'
+import { SetupDispatch } from '../../setup/store'
+import { PlotsDispatch } from '../../plots/store'
 
 const signalInitialized = () =>
   sendMessage({ type: MessageFromWebviewType.INITIALIZED })
 
 export function useVsCodeMessaging<T extends WebviewData>(
-  handler?: (event: { data: MessageToWebview<T> }) => void
+  feedStore: (
+    data: MessageToWebview<T>,
+    dispatch: PlotsDispatch | ExperimentsDispatch | SetupDispatch
+  ) => void
 ) {
+  const dispatch = useDispatch()
+
+  const handler = useCallback(
+    ({ data }: { data: MessageToWebview<T> }) => {
+      feedStore(data, dispatch)
+    },
+    [dispatch, feedStore]
+  )
+
   useEffect(() => {
     signalInitialized()
   }, [])
